requirements.txt
langchain
streamlit
openai
gtts
pygame
requests
python-dotenv